+++ src/co/com/samtel/entity/sql/BigRecuperaCarteraCastigada.java
@@ -47,7 +47,7 @@ public class BigRecuperaCarteraCastigada implements Serializable {
 	private BigDecimal f01morcor;
 	
 	@Column(name = "m_accfun")
-	private Integer f01accfun;
+	private Short f01accfun;
 		
 	@Column(name = "m_mipyme")
 	private BigDecimal f01mipyme;	
@@ -154,13 +154,13 @@ public class BigRecuperaCarteraCastigada implements Serializable {
 
 	public void setF01morcor(BigDecimal f01morcor) {
 		this.f01morcor = f01morcor;
-	}
+	}	
 
-	public Integer getF01accfun() {
+	public Short getF01accfun() {
 		return f01accfun;
 	}
 
-	public void setF01accfun(Integer f01accfun) {
+	public void setF01accfun(Short f01accfun) {
 		this.f01accfun = f01accfun;
 	}
